Extract to configFile subfolder, in order to clarify probable modularity for later extraction to a separate library used by multiple idproviders (#946)
