+++ src/main/java/com/enonic/app/oidcidprovider/lib/configFile/AbstractIdProviderHandler.java
@@ -1,4 +1,4 @@
-package com.enonic.app.oidcidprovider.lib.auth;
+package com.enonic.app.oidcidprovider.lib.configFile;
 
 import com.enonic.xp.script.bean.BeanContext;
 import com.enonic.xp.script.bean.ScriptBean;
