+++ src/main/java/com/enonic/app/oidcidprovider/lib/configFile/IdProviderConfigMapper.java
@@ -1,4 +1,4 @@
-package com.enonic.app.oidcidprovider.lib.auth;
+package com.enonic.app.oidcidprovider.lib.configFile;
 
 import com.enonic.xp.data.Property;
 import com.enonic.xp.data.PropertyArray;
