+++ src/main/java/com/enonic/app/oidcidprovider/lib/configFile/ScriptValueToIdProviderAccessControlListTranslator.java
@@ -1,4 +1,4 @@
-package com.enonic.app.oidcidprovider.lib.auth;
+package com.enonic.app.oidcidprovider.lib.configFile;
 
 import java.util.List;
 import java.util.Objects;
