+++ src/main/resources/lib/configFile/configFile.js
